Quarks: Never hide history view
If we hide the view with setVisibility(), we don't let animations
finish and never free the views. If the list has no entries, it's
not visible to the user, so just omit the call to setVisibility().
Note that this change allows RecyclerView.getChildAt(0) to return
null, so add the needed null pointer check.

Change-Id: Ieb0520122de1b4e51f72c4d4e5392947d1186caf
Signed-off-by: Joe Maples <joe@frap129.org>