Quarks: Remove the SwipeRefreshLayout
This is just causing issues. Some webpages make the SwipeRefreshLayout
intercept all the downwards swipes, triggering an unwanted page
refresh rather than scrolling the content of the page. This appears
to happen with pages that have a computeVerticalScrollRange() equals
to their getHeight(). However checking these two values is not enough
for some pages that generate content dynamically. Let's just get rid
of the SwipeRefreshLayout and all the related hacks.

Change-Id: I172cdc310cd86d1cd0d0f400930bd41c2da34818
Signed-off-by: CertifiedBlyndGuy <certifiedblyndguy@gmail.com>